refactor(ais-instant-search): move from pure factory function to augmented factory
This limits the amounts of conditionals in the code
